Fix ci tests
